Split provider implementations in separate files; Add Gemini
